GCP/flask_skeleton/requirements.txt
Flask==2.2.3 
Werkzeug==2.3.7
gunicorn==20.1.0
requests==2.28.2
google-cloud==0.34.0
google-cloud-pubsub==2.19